requirements.txt
# Core API
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
sqlalchemy==2.0.25

# Vision & ML
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
ultralytics==8.1.0  # YOLOv8
numpy==1.26.3

# Data processing
pandas==2.2.0
openpyxl==3.1.2  # For Excel files
requests==2.31.0

# Search & matching
rapidfuzz==3.6.1  # Fast fuzzy string matching

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
httpx==0.26.0
